Update for `abstract-level`
Moves many modules (but not yet all) to the Legacy section and adds
new `abstract-level` based modules.

Also adds a `timeless` option for modules, to make the "last commit"
badge green, to indicate that the module is feature-complete and can
be used even if it seems old by lack of activity.